deploy: Fix creation with`:latest` tag and CI
If a new deployment is deployed it previously created it with the 
non-CI_SHA1 version. This can lead to problems if :latest does exist
or could cause problems. If :latest is used, it could be from a non-prod
branch, or test code that doesn’t make sense for the new deployment to
the namespace.

This moves the CI_SHA1 detection and deployment file up before attempting
to create/apply the deployment