OPTIONAL_CONFIGS: Allow missing files unless deploying
This fixes a case that prevented commands from running when using S3/sops secrets without the use of ROK8S_ENABLE_OPTIONAL_CONFIGS.

The pattern for s3/sops secrets is to define a file in `S3_SECRETS` or `SOPS_SECRETS` as well as in `SECRETS`. The likely case with this setup is that the `SECRETS` file wouldn't be created until after running `k8s-secrets-from-s3` or `k8s-sops-secret-decrypt`. This aren't runnable though because the config reading can't find the file in `SECRETS`.

This change makes it so that the `ROK8S_ENABLE_OPTIONAL_CONFIGS` is only used for the deploy commands, allowing other commands to run even if files are missing (the previous default).