Added functionality for hiding secrets in CD
Be default the `kubectl apply` command will output the
object it is trying to apply in stderr. This currently
will expose secrets to the CI/CD output in any env if the
secret data is invalid.

This change attempts to scrape any potentially sensitive
information by assuming kubectl will keep the objects in
one line which also contains "kind":"Secret".

The downside to this is obviously if kubectl changes its
convention then we will have to adjust this method. A
different approach could include trying to collect
keywords from the file being applied to be able to more
confidently redact information.